requirements.txt
amqp==5.0.6
Babel==2.8.0
billiard==3.6.3.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-And-Redis==1.0.0
Flask-Babel==2.0.0
Flask-Cors==3.0.8
Flask-Script==2.0.6
flask-expects-json==1.5.0
flower==0.9.4
gevent==21.1.2
greenlet==1.0.0
httplib2==0.18.1
humanize==0.5.1
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
msgpack==1.0.0
PyJWT==2.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.16.0
python-slugify==4.0.0
pytz==2020.1
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.25.1
rsa==4.6
sentry-sdk==0.10.2
six==1.15.0
text-unidecode==1.3
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
Werkzeug==1.0.0
zipp==3.3.0
cryptography==3.3.1
gunicorn==20.1.0
bcrypt==3.2.0
pymongo==3.11.3
jsonschema~=3.2.0
pymodm==0.4.3
marshmallow==3.13.0
# uWSGI==2.0.17.1
kafka-python==2.0.2
apscheduler==3.9.1
boto3==1.21.38
websockets==10.3
pydash==5.1.0
thor_requests==1.2.2

